# Performance `_seriesIteratorGetNext`: Remove unecessary free/alloc on end of each chunk 