--- a/src/chunk.c
+++ b/src/chunk.c
@@ -172,22 +172,28 @@ size_t Uncompressed_DelRange(Chunk_t *chunk, timestamp_t startTs, timestamp_t en
     return deleted_count;
 }
 
-ChunkIter_t *Uncompressed_NewChunkIterator(Chunk_t *chunk,
-                                           int options,
-                                           ChunkIterFuncs *retChunkIterClass) {
-    ChunkIterator *iter = (ChunkIterator *)calloc(1, sizeof(ChunkIterator));
+void Uncompressed_ResetChunkIterator(ChunkIter_t *iterator,
+                                     Chunk_t *chunk,
+                                     int options,
+                                     ChunkIterFuncs *retChunkIterClass) {
+    ChunkIterator *iter = (ChunkIterator *)iterator;
     iter->chunk = chunk;
-    iter->options = options;
     if (options & CHUNK_ITER_OP_REVERSE) { // iterate from last to first
         iter->currentIndex = iter->chunk->num_samples - 1;
     } else { // iterate from first to last
         iter->currentIndex = 0;
     }
-
     if (retChunkIterClass != NULL) {
         *retChunkIterClass = *GetChunkIteratorClass(CHUNK_REGULAR);
     }
+}
 
+ChunkIter_t *Uncompressed_NewChunkIterator(Chunk_t *chunk,
+                                           int options,
+                                           ChunkIterFuncs *retChunkIterClass) {
+    ChunkIterator *iter = (ChunkIterator *)calloc(1, sizeof(ChunkIterator));
+    iter->options = options;
+    Uncompressed_ResetChunkIterator(iter, chunk, options, retChunkIterClass);
     return (ChunkIter_t *)iter;
 }