--- a/src/compressed_chunk.h
+++ b/src/compressed_chunk.h
@@ -28,10 +28,7 @@ size_t Compressed_DelRange(Chunk_t *chunk, timestamp_t startTs, timestamp_t endT
 ChunkIter_t *Compressed_NewChunkIterator(Chunk_t *chunk,
                                          int options,
                                          ChunkIterFuncs *retChunkIterClass);
-void Compressed_ResetChunkIterator(ChunkIter_t *iterator,
-                                   Chunk_t *chunk,
-                                   int options,
-                                   ChunkIterFuncs *retChunkIterClass);
+void Compressed_ResetChunkIterator(ChunkIter_t *iterator, Chunk_t *chunk);
 ChunkResult Compressed_ChunkIteratorGetNext(ChunkIter_t *iter, Sample *sample);
 void Compressed_FreeChunkIterator(ChunkIter_t *iter);