--- a/src/generic_chunk.c
+++ b/src/generic_chunk.c
@@ -16,6 +16,7 @@ static ChunkFuncs regChunk = {
     .DelRange = Uncompressed_DelRange,
 
     .NewChunkIterator = Uncompressed_NewChunkIterator,
+    .ResetChunkIterator = Uncompressed_ResetChunkIterator,
 
     .GetChunkSize = Uncompressed_GetChunkSize,
     .GetNumOfSample = Uncompressed_NumOfSample,
@@ -45,6 +46,7 @@ static ChunkFuncs comprChunk = {
     .DelRange = Compressed_DelRange,
 
     .NewChunkIterator = Compressed_NewChunkIterator,
+    .ResetChunkIterator = Compressed_ResetChunkIterator,
 
     .GetChunkSize = Compressed_GetChunkSize,
     .GetNumOfSample = Compressed_ChunkNumOfSample,