--- a/src/generic_chunk.h
+++ b/src/generic_chunk.h
@@ -67,6 +67,10 @@ typedef struct ChunkFuncs
     ChunkIter_t *(*NewChunkIterator)(Chunk_t *chunk,
                                      int options,
                                      ChunkIterFuncs *retChunkIterClass);
+    void (*ResetChunkIterator)(ChunkIter_t *iterator,
+                               Chunk_t *chunk,
+                               int options,
+                               ChunkIterFuncs *retChunkIterClass);
 
     size_t (*GetChunkSize)(Chunk_t *chunk, bool includeStruct);
     u_int64_t (*GetNumOfSample)(Chunk_t *chunk);