This PR removes some unecessary work on `_seriesIteratorGetNext`. We've seen on https://github.com/RedisTimeSeries/RedisTimeSeries/issues/753 and https://github.com/RedisTimeSeries/RedisTimeSeries/issues/747 that  `_seriesIteratorGetNext` is on the top2/3 hotspots. 
We should further iterate on this hotspot prior considering the overhead of 17.56% minimized! ( more PRs to come :) )

Overall impact on standalone numbers:
- `cpu-max-all-1`: from **1085.14 queries/sec** (p50=56.22 ms) to **1104.53 queries/sec** (p50=55.22 ms)
- `double-groupby-5`: from 19.86 queries/sec **(p50=3236.22 ms)** to 20.06 queries/sec **(p50=3184.00 ms)**

To trigger benchmarks again:
```
make clean build benchmark BENCHMARK_REPETITIONS=3 BENCHMARK=tsbs-scale100-cpu-max-all-1.yml
make clean build benchmark BENCHMARK_REPETITIONS=3 BENCHMARK=tsbs-scale100-double-groupby-5.yml

```


## master

```
# cpu-max-all-1:

Run complete after 10000 queries with 64 workers (Overall query rate 1085.14 queries/sec):
RedisTimeSeries max of all CPU metrics, random    1 hosts, random 8h0m0s by 1h:
min:     1.47ms, med:    56.22ms, mean:    58.85ms, max:  149.77ms, stddev:    11.83ms, sum: 588.5sec, count: 10000

# double-groupby-5:
Run complete after 1000 queries with 64 workers (Overall query rate 19.86 queries/sec):
RedisTimeSeries mean of 5 metrics, all hosts, random 12h0m0s by 1h:
min:    48.86ms, med:  3236.22ms, mean:  3181.02ms, max: 4448.26ms, stddev:   500.46ms, sum: 3181.0sec, count: 1000
```

## branch:
```
# cpu-max-all-1:

Run complete after 10000 queries with 64 workers (Overall query rate 1104.53 queries/sec):
RedisTimeSeries max of all CPU metrics, random    1 hosts, random 8h0m0s by 1h:
min:     1.09ms, med:    55.22ms, mean:    57.75ms, max:  130.25ms, stddev:    11.91ms, sum: 577.5sec, count: 10000

# double-groupby-5:
Run complete after 1000 queries with 64 workers (Overall query rate 20.06 queries/sec):
RedisTimeSeries mean of 5 metrics, all hosts, random 12h0m0s by 1h:
min:    49.49ms, med:  3184.00ms, mean:  3147.11ms, max: 4370.18ms, stddev:   494.50ms, sum: 3147.1sec, count: 1000
```